% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov1.mxPBF.R
\name{cov1.mxPBF}
\alias{cov1.mxPBF}
\title{One-Sample Covariance Test with Maximum Pairwise Bayes Factor}
\usage{
cov1.mxPBF(X, Sigma0 = diag(ncol(X)), a0 = 2, b0 = 2, gamma = 1,
  nthreads = 1)
}
\arguments{
\item{X}{an \eqn{(n\times p)} data matrix where each row is an observation.}

\item{Sigma0}{a \eqn{(p\times p)} given covariance matrix.}

\item{a0}{shape parameter for inverse-gamma prior.}

\item{b0}{scale parameter for inverse-gamma prior.}

\item{gamma}{non-negative variance scaling parameter.}

\item{nthreads}{number of threads for parallel execution via OpenMP.}
}
\value{
a (list) object of \code{S3} class \code{htest} containing: \describe{
\item{statistic}{maximum of pairwise Bayes factor.}
\item{alternative}{alternative hypothesis.}
\item{method}{name of the test.}
\item{data.name}{name(s) of provided sample data.}
\item{log.BF.mat}{matrix of pairwise Bayes factors in natural log.}
}
}
\description{
Not Written Here - No Reference Yet.
}
\examples{
\dontrun{
## generate from multivariate normal with identity covariance.
data = matrix(rnorm(100*5), ncol=10)

## run test with different parameters
out1 = cov1.mxPBF(data)
out2 = cov1.mxPBF(data, a0=5.0, b0=5.0) # change some params

## visualize two Bayes Factor matrices
par(mfrow=c(1,2), pty="s")
image(exp(out1$log.BF.mat)[,10:1], main="default")
image(exp(out2$log.BF.mat)[,10:1], main="a0=b0=5.0")
}

}
