% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2.mxPBF.R
\name{cov2.mxPBF}
\alias{cov2.mxPBF}
\title{Two-Sample Covariance Test with Maximum Pairwise Bayes Factor}
\usage{
cov2.mxPBF(X, Y, a0 = 2, b0 = 2, gamma = 1, nCores = 1)
}
\description{
Two-Sample Covariance Test with Maximum Pairwise Bayes Factor
}
