% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covk.2007Schott.R
\name{covk.2007Schott}
\alias{covk.2007Schott}
\title{Test for Homogeneity of Variance by Schott (2007)}
\usage{
covk.2007Schott(dlist)
}
\description{
Test for Homogeneity of Variance by Schott (2007)
}
\examples{
\donttest{
### generate k-sample from multivariate normal distribution
k = 5            # set 5-sample case
n = 100          # 100 observations in R^5
p = 5
mydata = list()
for (i in 1:k){
  mydata[[i]] = matrix(rnorm(n*p), ncol=p)
}

}
}
