% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqdist.2014BG.R
\name{eqdist.2014BG}
\alias{eqdist.2014BG}
\title{Test for Equality of Two Distributions by Biswas and Ghosh (2014)}
\usage{
eqdist.2014BG(X, Y, method = c("asymptotic", "permutation"),
  nreps = 2000)
}
\description{
Test for Equality of Two Distributions by Biswas and Ghosh (2014)
}
\examples{
\dontrun{
## compare asymptotic and permutation-based powers
set.seed(777)
ntest  = 100 
pval.a = rep(0,ntest)
pval.p = rep(0,ntest)

for (i in 1:ntest){
  x = rnorm(100)
  y = rnorm(100)
  
  pval.a[i] = ifelse(eqdist.2014BG(x,y,method="a")$p.value<0.05,1,0)
  pval.p[i] = ifelse(eqdist.2014BG(x,y,method="p",nreps=100)$p.value <0.05,1,0)
}

## print the result
cat(paste(
sprintf("\\n* EMPIRICAL TYPE 1 ERROR COMPARISON \\n"),
sprintf("* Asymptotics :  \%.3f\\n", sum(pval.a/ntest)),
sprintf("* Permutation :  \%.3f\\n", sum(pval.p/ntest)),sep=""))
}

}
\references{
\insertRef{biswas_nonparametric_2014}{SHT}
}
