% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean1.Srivastava2008.R
\name{mean1.Srivastava2008}
\alias{mean1.Srivastava2008}
\title{One-Sample Test for High-Dimensional Mean by Srivastava and Du (2008)}
\usage{
mean1.Srivastava2008(X, mu0 = rep(0, ncol(X)), alpha = 0.05)
}
\description{
One-Sample Test for High-Dimensional Mean by Srivastava and Du (2008)
}
\references{
\insertRef{srivastava_test_2008}{SHT}
}
\author{
Kisung You
}
