% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean1.ttest.R
\name{mean1.ttest}
\alias{mean1.ttest}
\title{One-Sample Student's t-test for Univariate Mean}
\usage{
mean1.ttest(x, mu0 = 0, alternative = c("two.sided", "less",
  "greater"), alpha = 0.05)
}
\arguments{
\item{x}{a length-\eqn{n} data vector.}

\item{mu0}{hypothesized variance \eqn{\sigma_0^2}.}

\item{alternative}{specifying the alternative hypothesis.}

\item{alpha}{significance level.}
}
\value{
a (list) object of \code{S3} class \code{hypothesis} containing: \describe{
\item{method}{name of the test.}
\item{statistic}{a test statistic.}
\item{p.value}{\eqn{p}-value under current setting.}
\item{significance}{a user-specified significance level.}
\item{alternative}{alternative hypothesis.}
\item{conclusion}{conclusion by \eqn{p}-value decision rule.}
}
}
\description{
Given an univariate sample \eqn{x}, it tests
\deqn{H_0 : \mu_x = \mu_0\quad vs\quad H_1 : \mu_x \neq \mu_0}
using the procedure by Student (1908).
}
\examples{
\donttest{
## empirical Type 1 error 
niter   = 1000
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  x = rnorm(10)         # sample from N(0,1)
  counter[i] = ifelse(mean1.ttest(x)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\\n* Example for 'mean1.ttest'\\n\\n",
sprintf("* number of rejections   : \%d\\n",sum(counter)),
sprintf("* total number of trials : \%d\\n",niter),
sprintf("* empirical Type 1 error : \%.4f\\n", sum(counter/niter)),sep=""))
}

}
\references{
\insertRef{student_probable_1908}{SHT}

\insertRef{student_probable_1908-1}{SHT}
}
