% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.1958Dempster.R
\name{mean2.1958Dempster}
\alias{mean2.1958Dempster}
\title{Two-Sample Test for High-Dimensional Means by Dempster (1958, 1960)}
\usage{
mean2.1958Dempster(X, Y, alpha = 0.05)
}
\arguments{
\item{X}{an \eqn{(n_x \times p)} data matrix of 1st sample.}

\item{Y}{an \eqn{(n_y \times p)} data matrix of 2nd sample.}

\item{alpha}{significance level.}
}
\value{
a (list) object of \code{S3} class \code{hypothesis} containing: \describe{
\item{method}{name of the test.}
\item{statistic}{a test statistic.}
\item{p.value}{\eqn{p}-value under current setting.}
\item{significance}{a user-specified significance level.}
\item{alternative}{alternative hypothesis.}
\item{conclusion}{conclusion by \eqn{p}-value decision rule.}
}
}
\description{
Given two multivariate data \eqn{X} and \eqn{Y} of same dimension, it tests
\deqn{H_0 : \mu_x = \mu_y\quad vs\quad H_1 : \mu_x \neq \mu_y}
using the procedure by Dempster (1958, 1960).
}
\examples{
\donttest{
## empirical Type 1 error 
niter   = 1000
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  X = matrix(rnorm(50*5), ncol=10)
  Y = matrix(rnorm(50*5), ncol=10)
  
  counter[i] = ifelse(mean2.1958Dempster(X,Y)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\\n* Example for 'mean2.1958Dempster'\\n\\n",
sprintf("* number of rejections   : \%d\\n",sum(counter)),
sprintf("* total number of trials : \%d\\n",niter),
sprintf("* empirical Type 1 error : \%.4f\\n", sum(counter/niter)),sep=""))
}

}
\references{
\insertRef{dempster_high_1958}{SHT}

\insertRef{dempster_significance_1960}{SHT}
}
\author{
Kisung You
}
