% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.Bai1996.R
\name{mean2.Bai1996}
\alias{mean2.Bai1996}
\title{Two-Sample Test for High-Dimensional Means by Bai and Saranadasa (1996)}
\usage{
mean2.Bai1996(X, Y, alpha = 0.05)
}
\description{
Two-Sample Test for High-Dimensional Means by Bai and Saranadasa (1996)
}
\references{
\insertRef{bai_high_1996}{SHT}
}
\author{
Kisung You
}
