% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.Cai2014.R
\name{mean2.Cai2014}
\alias{mean2.Cai2014}
\title{Two-Sample Test for High-Dimensional Means by Cai et al (2014)}
\usage{
mean2.Cai2014(X, Y, alpha = 0.05)
}
\description{
Two-Sample Test for High-Dimensional Means by Cai et al (2014)
}
\references{
\insertRef{cai_two-sample_2014}{SHT}
}
\author{
Kisung You
}
