% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.Chen2010.R
\name{mean2.Chen2010}
\alias{mean2.Chen2010}
\title{Two-Sample Test for High-Dimensional Means by Chen and Qin (2014)}
\usage{
mean2.Chen2010(X, Y, alpha = 0.05)
}
\description{
Two-Sample Test for High-Dimensional Means by Chen and Qin (2014)
}
\references{
\insertRef{chen_two-sample_2010}{SHT}
}
\author{
Kisung You
}
