% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.Srivastava2008.R
\name{mean2.Srivastava2008}
\alias{mean2.Srivastava2008}
\title{Two-Sample Test for High-Dimensional Means by Srivastava and Du (2008)}
\usage{
mean2.Srivastava2008(X, Y, alpha = 0.05)
}
\description{
Two-Sample Test for High-Dimensional Means by Srivastava and Du (2008)
}
\references{
\insertRef{srivastava_test_2008}{SHT}
}
\author{
Kisung You
}
