% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.1965SW.R
\name{norm.1965SW}
\alias{norm.1965SW}
\title{Univariate Test of Normality by Shapiro and Wilk (1965)}
\usage{
norm.1965SW(x)
}
\description{
Univariate Test of Normality by Shapiro and Wilk (1965)
}
\examples{
\donttest{
## generate samples from several distributions
x = stats::runif(496)            # uniform
y = stats::rgamma(496, shape=2)  # gamma
z = stats::rlnorm(496)           # log-normal

## test above samples
test.x = norm.1965SW(x) # uniform
test.y = norm.1965SW(y) # gamma
test.z = norm.1965SW(z) # log-normal
}

}
