% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.1972SF.R
\name{norm.1972SF}
\alias{norm.1972SF}
\title{Univariate Test of Normality by Shapiro and Francia (1972)}
\usage{
norm.1972SF(x)
}
\description{
Univariate Test of Normality by Shapiro and Francia (1972)
}
\examples{
\donttest{
## generate samples from several distributions
x = stats::runif(496)            # uniform
y = stats::rgamma(496, shape=2)  # gamma
z = stats::rlnorm(496)           # log-normal

## test above samples
test.x = norm.1972SF(x) # uniform
test.y = norm.1972SF(y) # gamma
test.z = norm.1972SF(z) # log-normal
}

}
