% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.1980JB.R
\name{norm.1980JB}
\alias{norm.1980JB}
\title{Univariate Test of Normality by Jarque and Bera (1980)}
\usage{
norm.1980JB(x, method = c("asymptotic", "MC"), nreps = 2000)
}
\description{
Univariate Test of Normality by Jarque and Bera (1980)
}
\examples{
\donttest{
## generate samples from uniform distribution
x = runif(496)

## test with both methods of attaining p-values
test1 = norm.1980JB(x, method="a") # Asymptotics
test2 = norm.1980JB(x, method="m") # Monte Carlo 
}

}
