% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.1996AJB.R
\name{norm.1996AJB}
\alias{norm.1996AJB}
\title{Adjusted Jarque-Bera Test of Univariate Normality by Urzua (1996)}
\usage{
norm.1996AJB(x, method = c("asymptotic", "MC"), nreps = 2000)
}
\description{
Adjusted Jarque-Bera Test of Univariate Normality by Urzua (1996)
}
\examples{
\donttest{
## generate samples from uniform distribution
x = runif(496)

## test with both methods of attaining p-values
test1 = norm.1996AJB(x, method="a") # Asymptotics
test2 = norm.1996AJB(x, method="m") # Monte Carlo 
}

}
