% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.2008RJB.R
\name{norm.2008RJB}
\alias{norm.2008RJB}
\title{Robust Jarque-Bera Test of Univariate Normality by Gel and Gastwirth (2008)}
\usage{
norm.2008RJB(x, C1 = 6, C2 = 24, method = c("asymptotic", "MC"),
  nreps = 2000)
}
\description{
Robust Jarque-Bera Test of Univariate Normality by Gel and Gastwirth (2008)
}
\examples{
\donttest{
## generate samples from uniform distribution
x = runif(496)

## test with both methods of attaining p-values
test1 = norm.2008RJB(x, method="a") # Asymptotics
test2 = norm.2008RJB(x, method="m") # Monte Carlo 
}

}
