% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-SHT.R
\docType{package}
\name{SHT}
\alias{SHT}
\alias{SHT-package}
\title{Statistical Hypothesis Testing Toolbox}
\description{
Testing statistical hypotheses is one of the most significant inference problems in statistics. 
As time goes by, the community has witnessed surge of unorthodox settings such as high-dimensionality 
as well as atypical problems galore. We, though not complete nor ambitious, try to gather up 
some of frequently appearing tests as many as possible. Entire list of available tests 
can be seen at README file.
}
