% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silent_sim1.2020YLL.R
\name{sim1.2020YLL}
\alias{sim1.2020YLL}
\title{Temporary : Random Projection for One-Sample Simultaneous Testing}
\usage{
sim1.2020YLL(X, mu0 = rep(0, ncol(X)), Sigma0 = diag(ncol(X)), m = 50)
}
\description{
Temporary : Random Projection for One-Sample Simultaneous Testing
}
\examples{
## CRAN-purpose small example
smallX = matrix(rnorm(10*3),ncol=3)
sim1.2020YLL(smallX) # run the test

\dontrun{
## empirical Type 1 error 
niter   = 1000
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  X = matrix(rnorm(50*10), ncol=10)
  counter[i] = ifelse(sim1.2020YLL(X)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\n* Example for 'sim1.2020YLL'\n","*\n",
"* number of rejections   : ", sum(counter),"\n",
"* total number of trials : ", niter,"\n",
"* empirical Type 1 error : ",round(sum(counter/niter),5),"\n",sep=""))
}

}
