% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var1.chisq.R
\name{var1.chisq}
\alias{var1.chisq}
\title{One-Sample Chi-Square Test for Variance}
\usage{
var1.chisq(x, var0 = 1, alternative = c("two.sided", "less", "greater"),
  alpha = 0.05)
}
\description{
One-Sample Chi-Square Test for Variance
}
