% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vark.1974Brown.R
\name{vark.1974Brown}
\alias{vark.1974Brown}
\title{Brown-Forsythe Test for Homogeneity of Variance}
\usage{
vark.1974Brown(dlist, alpha = 0.05)
}
\arguments{
\item{dlist}{a list of length \eqn{k} where each element is a sample vector.}

\item{alpha}{significance level.}
}
\value{
a (list) object of \code{S3} class \code{hypothesis} containing: \describe{
\item{method}{name of the test.}
\item{statistic}{a test statistic.}
\item{p.value}{\eqn{p}-value under current setting.}
\item{significance}{a user-specified significance level.}
\item{alternative}{alternative hypothesis.}
\item{conclusion}{conclusion by \eqn{p}-value decision rule.}
}
}
\description{
Given univariate samples \eqn{X_1~,\ldots,~X_k}, it tests
\deqn{H_0 : \sigma_1^2 = \cdots \sigma_k^2\quad vs\quad H_1 : \textrm{at least one equality does not hold}}
using the procedure by Brown and Forsythe (1974).
}
\examples{
\donttest{
## test when k=5 (samples)
## empirical Type 1 error 
niter   = 1000
counter = rep(0,niter)  # record p-values
for (i in 1:niter){
  mylist = list()
  for (j in 1:5){
     mylist[[j]] = rnorm(50)   
  }
  
  counter[i] = ifelse(vark.1974Brown(mylist)$p.value < 0.05, 1, 0)
}

## print the result
cat(paste("\\n* Example for 'vark.1974Brown'\\n\\n",
sprintf("* number of rejections   : \%d\\n",sum(counter)),
sprintf("* total number of trials : \%d\\n",niter),
sprintf("* empirical Type 1 error : \%.4f\\n", sum(counter/niter)),sep=""))
}

}
\references{
\insertRef{levene_robust_1960}{SHT}
}
