% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vark.BrownForsythe.R
\name{vark.BrownForsythe}
\alias{vark.BrownForsythe}
\title{Brown-Forsythe Test for Homogeneity of Variance}
\usage{
vark.BrownForsythe(dlist, alpha = 0.05)
}
\description{
Brown-Forsythe Test for Homogeneity of Variance
}
