% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean1.Hotelling.R
\name{mean1.Hotelling}
\alias{mean1.Hotelling}
\title{One-Sample Hotelling's T-squared Test for Multivariate Mean}
\usage{
mean1.Hotelling(X, mu0 = rep(0, ncol(X)), alpha = 0.05)
}
\description{
One-Sample Hotelling's T-squared Test for Multivariate Mean
}
