% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean2.Student.R
\name{mean2.Student}
\alias{mean2.Student}
\title{Two-Sample Student's t-test for Univariate Mean}
\usage{
mean2.Student(x, y, alternative = c("two.sided", "less", "greater"),
  alpha = 0.05, paired = FALSE, var.equal = FALSE)
}
\arguments{
\item{x}{a vector of length \eqn{m}.}

\item{y}{a vector of length \eqn{n}.}

\item{alpha}{significance level, default set as 0.05.}

\item{paired}{a logical indicating to use a paired \eqn{t}-test.}

\item{var.equal}{a logical indicating to use Welch's variant.}
}
\description{
Two-Sample Student's t-test for Univariate Mean
}
