% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplex_unif.R
\name{simplex.uniform}
\alias{simplex.uniform}
\title{Classical Test of Uniformity}
\usage{
simplex.uniform(X, method)
}
\arguments{
\item{X}{an \eqn{(n\times p)} data matrix where each row is an observation.}

\item{method}{name of the method to be used, including\describe{
\item{LRT}{likelihood-ratio test with the Dirichlet distribution.}
\item{LRTsym}{likelihood-ratio test using the symmetric Dirichlet distribution (default).}
}}
}
\description{
Classical Test of Uniformity
}
\concept{simplex}
